# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tasks Management API",
        "description": "This is an example API for tasks management",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/",
            "description": "Base path for all API endpoints"
        }
    ],
    "paths": {
...
        "/posts": {
            "get": {
                "tags": [
                    "Post Management"
                ],
                "summary": "Get a paginated list of posts",
                "description": "Retrieves a list of posts owned by the authenticated user. Use query parameters for filtering and pagination.",
                "operationId": "cce7863441436fd49dc474491f658292",
                "parameters": [
                    {
                        "name": "filters",
                        "in": "query",
                        "description": "Optional filters for searching posts (refer to your specific implementation)",
                        "schema": {
                            "type": "array",
                            "items": {
                                "description": "Filter posts by their status",
                                "type": "string",
                                "enum": [
                                    "completed",
                                    "create",
                                    "system_completed"
                                ]
                            }
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "description": "Number of items per page (default 10)",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "items": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/PostSchema"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctumAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Post Management"
                ],
                "summary": "Create a new post",
                "description": "Creates a new post with the provided details.",
                "operationId": "21a6ac5f9280a3cf958c3d066b5600ee",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreatePostSchema"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Post created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/PostSchema"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                }
            }
        },
        "/posts/{postId}": {
            "get": {
                "tags": [
                    "Post Management"
                ],
                "summary": "Get a post by ID",
                "description": "Retrieves a single post identified by its ID.",
                "operationId": "04e28d03abbc1c8cbe8d01b92b4011b2",
                "parameters": [
                    {
                        "name": "postId",
                        "in": "path",
                        "description": "ID of the post",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PostSchema"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Post not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Post Management"
                ],
                "summary": "Update a post",
                "description": "Updates an existing post with the provided details.",
                "operationId": "26847e8ad19915583bd05e580ec8c6b0",
                "parameters": [
                    {
                        "name": "postId",
                        "in": "path",
                        "description": "ID of the post to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdatePostSchema"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Post updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/PostSchema"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden: Could not update this post"
                    },
                    "404": {
                        "description": "Post not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Post Management"
                ],
                "summary": "Delete a post",
                "description": "Deletes a post with the provided ID.",
                "operationId": "d477c7dc08966e54b1a008f3cf9c5439",
                "parameters": [
                    {
                        "name": "postId",
                        "in": "path",
                        "description": "ID of the post to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Post deleted successfully"
                    },
                    "404": {
                        "description": "Post not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "CreatePostSchema": {
                "properties": {
                    "content": {
                        "description": "post's content",
                        "type": "string"
                    },
                    "visibility": {
                        "description": "post's visibility",
                        "type": "string"
                    },
                    "media": {
                        "description": "post's media",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "UpdatePostSchema": {
                "properties": {
                    "content": {
                        "description": "post's content",
                        "type": "string"
                    },
                    "visibility": {
                        "description": "post's visibility",
                        "type": "string"
                    },
                    "media": {
                        "description": "post's media",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "PostSchema": {
                "properties": {
                    "id": {
                        "description": "Post ID",
                        "type": "integer"
                    },
                    "content": {
                        "description": "Post content",
                        "type": "string"
                    },
                    "visibility": {
                        "description": "Post visibility",
                        "type": "string"
                    },
                    "created at": {
                        "description": "Post created_at datetime",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Post updated_at datetime",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctumAuth": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}